requirements.txt
Flask==2.3.3
Werkzeug==3.0.2
python-dotenv==1.0.1
boto3==1.34.98
